manejo de errores y componentes
